add validation when deleting

+++ src/controllers/maker.controller.js
@@ -1,5 +1,7 @@
 import mongoose from 'mongoose';
 import Maker from '../models/maker.model.js';
+import Model from '../models/model.model.js';
+import Printer from '../models/printer.model.js';
 import { transformType } from '../utilities/transformType.js';
 
 export const getMakers = async (req, res) => {
@@ -68,6 +70,24 @@ export const deleteMaker = async (req, res) => {
       return res.status(400).json({ message: 'ID de fabricante inválido' });
     }
 
+    const hasModels = await Model.exists({ maker: id });
+
+    if (hasModels) {
+      return res.status(400).json({
+        message:
+          'No puedes eliminar esta marca porque contiene modelos asociados',
+      });
+    }
+
+    const isUsed = await isMakerInUse(id);
+
+    if (isUsed) {
+      return res.status(400).json({
+        message:
+          'No puedes eliminar este fabricante porque está en uso en alguna categoría',
+      });
+    }
+
     const maker = await Maker.findOneAndDelete({ _id: id, type });
     if (!maker)
       return res.status(404).json({ message: 'Fabricante no encontrado' });
@@ -112,3 +132,12 @@ export const updateMaker = async (req, res) => {
     return res.status(404).json({ error: error.message });
   }
 };
+
+async function isMakerInUse(id) {
+  const printerExists = await Printer.exists({ maker: id });
+  //const monitorExists = await Monitor.exists({ maker: id });
+  //const computerExists = await Computer.exists({ maker: id });
+  //const peripheralExists = await Peripheral.exists({ maker: id });
+
+  return printerExists; //|| monitorExists || computerExists || peripheralExists;
+}
